backend/requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0